Adds bar graph detection
Adds functions that detects on which pages bar graphs were found. This replaces the detection of jet color maps. The results are stored in the database and are visualized in the web app.

As the worker queuing system with rq is not available for windows, the prediction is currently done directly after the tweets have been retrieved. This is suboptimal as the prediction is quite slow, especially for long papers.